Module Pricing Finished with All 62 tests passed

+++ TradingDashboard.Core/Entities/MarketSnapshot.cs
@@ -5,5 +5,7 @@ using System.Timers;
 
 namespace TradingDashboard.Core.Entities
 {
-    public record struct MarketSnapshot(string Ticker, decimal SpotPrice, double ImpliedVolatility, double RiskFreeRate, DateTimeOffset Timestamp, string Source);
+    public record struct MarketSnapshot(string Ticker, double SpotPrice, double ImpliedVolatility, double RiskFreeRate, DateTimeOffset Timestamp, string Source);
+
 }
+
