+++ TradingDashboard.Core/Entities/Position.cs
@@ -10,10 +10,10 @@ namespace TradingDashboard.Core.Entities
         public Guid PositionId { get; init; }
         public Instrument Instrument { get; set; }
         public int NetQuantity { get; set; }
-        public decimal AverageEntryPrice { get; set; }
-        public decimal CurrentMarketValue { get; set; }
-        public decimal UnrealizedPnL { get; set; }
-        public decimal DailyPnL { get; set; }
+        public double AverageEntryPrice { get; set; }
+        public double CurrentMarketValue { get; set; }
+        public double UnrealizedPnL { get; set; }
+        public double DailyPnL { get; set; }
         public Greeks LastGreeks { get; set; }
         public DateTimeOffset LastUpdatedAt { get; set; }
 
@@ -21,7 +21,7 @@ namespace TradingDashboard.Core.Entities
         public bool IsLong => this.NetQuantity > 0;
         public bool IsFlat => this.NetQuantity == 0;
 
-        public void UpdateMarketValue(decimal newSpot)
+        public void UpdateMarketValue(double newSpot)
         {
             CurrentMarketValue = NetQuantity * newSpot;
             UnrealizedPnL = CurrentMarketValue - NetQuantity*AverageEntryPrice;
