+++ TradingDashboard.Core/Interfaces/IRiskEngine.cs
@@ -7,7 +7,7 @@ namespace TradingDashboard.Core.Interfaces
 {
     interface IRiskEngine
     {
-        Task<decimal> CalculatePnl(MarketSnapshot SnapShot);
+        Task<double> CalculatePnl(MarketSnapshot SnapShot);
         Task<Greeks> CalculateGreeks();
         Task<IEnumerable<Alert>> VerifyRiskLimits();
     }
