+++ TradingDashbord.Pricing/Strategy/BlackScholesPricingStrategy .cs
@@ -11,8 +11,8 @@ namespace TradingDashbord.Pricing.Strategy
     {
         public async Task<PricingResult> PriceAsync(Instrument instrument, MarketSnapshot snapshot, CancellationToken ct = default)
         {
-            IPricer<decimal> pricer = new VanillaOptionPricer(instrument.ProductType);
-            decimal theoreticalPrice = await pricer.CalculatePrice(instrument, snapshot);
+            IPricer<double> pricer = new VanillaOptionPricer(instrument.ProductType);
+            double theoreticalPrice = await pricer.CalculatePrice(instrument, snapshot);
             Greeks greeks = await pricer.CalculateGreeks(instrument, snapshot);
             PricingResult res  = new PricingResult(theoreticalPrice, greeks, "BlackScholes", null);
             return res;
